Create utils::io::file for file handling helpers
