+++ lib/src/utils/io/mod.rs
@@ -3,4 +3,5 @@
 
 //! Common utilities for things like networking and IO.
 
+pub mod file;
 pub mod progress;
